semantickernel/server/requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cryptography==41.0.4
dataclasses==0.6
diffusers==0.2.2
filelock==3.4.1
Flask==2.2.5
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.22.0
packaging==21.3
Pillow==10.0.1
pycparser==2.21
pyparsing==3.0.9
PyYAML==6.0
regex==2022.10.31
requests==2.31.0
sacremoses==0.0.53
six==1.16.0
tokenizers==0.13.3
torch==2.2.1
tqdm==4.64.1
transformers==4.39.3
typing_extensions==4.6.1
urllib3==1.26.18
Werkzeug==2.2.3
zipp==3.6.0
pydantic==2.4.2
sentencepiece==0.1.99
einops==0.7.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
accelerate==0.25.0
xformers

